app factory + modifications in myip.html
