+++ .gitignore
@@ -2,4 +2,5 @@ venv
 *.jpg
 __pycache__
 *.pyc
-platzi
+no_project
+.sh
