+++ app/config.py
@@ -1,0 +1,2 @@
+class Config:
+    SECRET_KEY = 'SUPER SECRET'
