+++ app/forms.py
@@ -1,0 +1,8 @@
+from wtforms import PasswordField, SubmitField, StringField
+from wtforms.validators import DataRequired
+from flask_wtf import FlaskForm
+
+class LoginForm(FlaskForm):
+    username = StringField('username', validators=[DataRequired()])
+    password = PasswordField('password', validators=[DataRequired()])
+    submit = SubmitField('enviar')
