--- flask_run.sh
@@ -1,5 +1,0 @@
-source venv/bin/activate
-export FLASK_APP=main.py
-export FLASK_DEBUG=1
-export FLASK_ENV=development
-echo $(flask run)
