--- flask_test.sh
@@ -1,2 +1,0 @@
-export FLASK_AAPP=main.py
-flask test
