--- templates/myip.html
@@ -1,58 +1,0 @@
-{% extends 'base.html' %}
-{% import 'macros.html' as macros %}
-{% import 'bootstrap/wtf.html' as wtf %}
-
-
-<html>
-<head>
-    <title>
-        {% block title %}
-        {{ super() }}
-        My IPs
-        {% endblock %}
-    </title>
-  
-</head>
-<body>
-    {% block content %}
-        {% if username %}z
-            <p>Welcome: #{{ username | capitalize }}</pe>
-        {% endif %}
-        
-        <h2>Tu Dirección IP es: {{ ip_user }}</h2>
-        <h2>To Do:</h2>
-        <ul>
-            {% for todo in todos %}
-            {{ macros.render_todos(todo) }}
-            {% endfor %}
-        </ul>
-    
-    {% block scripts %}
-    {{ super() }}
-    {% endblock %} 
-
-    {{ macros.init_msg(check) }}
-    <div class="container">
-        {{ wtf.quick_form(login_form) }}
-    </div>
-    {% for message in get_flashed_messages() %}
-    <div class="alert alert-success alert-dissmisible">
-        {{ message }}
-        <button class="close" type="button" data-dismiss="alert">&times;</button>
-    </div>
-    {% endfor %}
-
-    {% endblock %}
-
-</body>
-
-</html>
-<!--
-    - se crea archivo base.htrml que contendrá información por defecto del sitio.
-    - Se extiende la base.html a myip.html.
-    # Extender parece ser una manera de incorporar información preliminar a un template.
-    - Se extinde base.html a myip.html.
-    - Se generan blockes para <title> personalizado.
-    - se mueve css a plantilla base.html.
-    - Se crean Macros y se mueven algunas funcionalidades a este .html
--->
